requirements.txt
langchain==0.3.3
langchain_community==0.3.2
langchain_openai==0.2.2
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.39.0
